+++ app.js
@@ -15,15 +15,3 @@ app.use("/", mainRouter);
 
 // Levantando el Servidor en el puerto 3030
 app.listen(3030, () => console.log("Server running in port 3030"));
-
-// Ruta Créditos
-app.get("/creditos", (req, res) => {
-  res.send(
-    "Coded by Marcos DV <br> - github.com/marcosdellavecchia <br> - linkedin.com/in/marcos-dv <br> - twitter.com/marcosdv"
-  );
-});
-
-// Ruta 404
-app.get("*", (req, res) => {
-  res.status(404).send("404 not found. <br> ¡Houston, poseemos problemas!");
-});
