+++ routes/main.js
@@ -1,12 +1,14 @@
 const express = require("express");
-
-let router = express.Router();
+const router = express.Router();
+const mainController = require("../controllers/main");
 
 // Ruta Raíz / ➝ Home
-router.get("/", (req, res) => {
-  res.send(
-    "Ni Superman, Iron Man o La Mujer Maravilla son tan importantes cómo las y los Heroes de carne y hueso que encontrarás en este sitio. Esperamos que ellas y ellos te sirvan como inspiración para poder cumplir tus objetivos. Recuerda: ¡nunca pares de creer en ti!."
-  );
-});
+router.get("/", mainController.home);
+
+// Ruta Créditos
+router.get("/creditos", mainController.credits);
+
+// Ruta 404
+router.get("*", mainController.notFound);
 
 module.exports = router;
